+++ tests/PRedisCacheTest.php
@@ -166,6 +166,12 @@ class PRedisCacheTest extends AbstractCache
         $this->assertEquals($cacheSamePrefix->get($key), $val);
         $this->assertNull($cacheOtherPrefix->get($key));
     }
+    
+    #[\Override]
+    public function testArrayDepth(): void
+    {
+        parent::testArrayDepth();
+    }
 
     #[\Override]
     public static function tearDownAfterClass(): void